Updated news feed manually.
This happens automatically on the server. If you 'make publish' before
running

scripts/fetch_news
cp utils/newsfetcher/output/* source/_static/news/

you will overwrite the news on the server with the older news you had 
cached on your drive. Even so, this will be corrected in a few minutes when 
the cron job reruns on the server.