Increase main menu font size
The current main menu (HOME ABOUT DOWNLOAD ...) at the top of the page has a hard to read font on my machine (RedHat with both Firefox and Chrome). Some vertical lines almost disappear, i.e the M in DOCUMENTATION. Am I the only one experiencing this? I discovered that it is the combination of the font (Lato), the font weight (300) and the size (12px). Going to 13px size mostly fixes the problem (in main.css at line 304). I think 14px looks the best (subjectively, tested on my desktop machine only). The normal text on the page is Lato/300/16px.

PS: trying to change this in the bitbucket online editor. Lets see how well that goes ...